requirements.txt
google-cloud-bigquery>=3.0.0
google-cloud-firestore>=2.0.0
google-generativeai>=0.5.0
requests>=2.31.0
scrapingbee>=2.0.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
jinja2>=3.1.0
sendgrid>=6.10.0
python-dateutil>=2.8.0
tenacity>=8.2.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
streamlit>=1.30.0
fastapi>=0.100.0
uvicorn>=0.23.0
db-dtypes>=1.1.1
